LayoutTests/websocket/tests/url-with-fragment-expected.txt

Make sure handshake with URL with fragment components success.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

url=ws://127.0.0.1:8880/websocket/tests/echo-location#fragment
WebSocket is open
received:ws://127.0.0.1:8880/websocket/tests/echo-location
WebSocket is closed
PASS handshake_success is true
PASS ws_location is "ws://127.0.0.1:8880/websocket/tests/echo-location"
PASS successfullyParsed is true

TEST COMPLETE
